rename auth directory to authentication
