+++ devU-api/src/authentication/login.saml/login.saml.controller.ts
@@ -4,7 +4,7 @@ import { User } from 'devu-shared-modules'
 import environment from '../../environment'
 
 import UserService from '../../entities/user/user.service'
-import AuthService from '../../auth/auth.service'
+import AuthService from '../auth.service'
 
 import { samlStrategy } from '../../utils/passport/saml.passport'
 import { refreshCookieOptions } from '../../utils/cookie.utils'
