+++ devU-api/src/router/index.ts
@@ -6,8 +6,8 @@ import swagger from '../utils/swagger.utils'
 import userCourse from '../entities/userCourse/userCourse.router'
 import assignments from '../entities/assignment/assignment.router'
 import courses from '../entities/course/course.router'
-import login from '../auth/login/login.router'
-import logout from '../auth/logout/logout.router'
+import login from '../authentication/login/login.router'
+import logout from '../authentication/logout/logout.router'
 import status from '../status/status.router'
 import submissions from '../entities/submission/submission.router'
 import users from '../entities/user/user.router'
@@ -22,7 +22,7 @@ import grader from '../entities/grader/grader.router'
 import categories from '../entities/category/category.router'
 import assignmentScore from '../entities/assignmentScore/assignmentScore.router'
 
-import { isAuthorized } from '../auth/auth.middleware'
+import { isAuthorized } from '../authentication/auth.middleware'
 
 import { NotFound } from '../utils/apiResponse.utils'
 import nonContainerAutoGraderRouter from "../entities/nonContainerAutoGrader/nonContainerAutoGrader.router";
